+++ src/test/java/org/folio/edge/sip2/repositories/LoginRepositoryTests.java
@@ -11,13 +11,9 @@ import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.when;
 
 import io.vertx.core.Future;
-import io.vertx.core.MultiMap;
 import io.vertx.core.Vertx;
-import io.vertx.core.impl.NoStackTraceThrowable;
-import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
-import java.util.Collections;
 import org.folio.edge.sip2.domain.messages.enumerations.PWDAlgorithm;
 import org.folio.edge.sip2.domain.messages.enumerations.UIDAlgorithm;
 import org.folio.edge.sip2.domain.messages.requests.Login;
@@ -103,7 +99,7 @@ public class LoginRepositoryTests {
   }
 
   @Test
-  public void canPatronLogin(Vertx vertx,
+  void canPatronLogin(Vertx vertx,
       VertxTestContext testContext,
       @Mock IResourceProvider<IRequestData> mockFolioProvider) {
     final String username = "test";
@@ -122,7 +118,7 @@ public class LoginRepositoryTests {
   }
 
   @Test
-  public void cannotPatronLogin(Vertx vertx,
+  void cannotPatronLogin(Vertx vertx,
       VertxTestContext testContext,
       @Mock IResourceProvider<IRequestData> mockFolioProvider) {
     final String username = "test";