+++ src/test/java/org/folio/edge/sip2/repositories/UsersRepositoryTests.java
@@ -1,6 +1,7 @@
 package org.folio.edge.sip2.repositories;
 
 import static org.folio.edge.sip2.api.support.TestUtils.getJsonFromFile;
+import static org.folio.util.PercentCodec.encode;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -19,7 +20,6 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
 import org.folio.edge.sip2.session.SessionData;
-import org.folio.edge.sip2.utils.Utils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -58,10 +58,6 @@ public class UsersRepositoryTests {
 
     final String barcode = "997383903573496";
     final String userId = "4f0e711c-d583-41e0-9555-b62f1725023f";
-    final String expectedUsersQueryPath = "/users?limit=1&query="
-        + Utils.encode("(barcode==" + barcode
-        + " or externalSystemId==" + barcode
-        + " or username==" + barcode + ')');
 
     final String expectedUsersBlQueryPath = "/bl-users/by-id/" + userId;
 
@@ -73,8 +69,7 @@ public class UsersRepositoryTests {
 
     doReturn(Future.succeededFuture(new FolioResource(userResponse,
         MultiMap.caseInsensitiveMultiMap().add("x-okapi-token", "1234"))))
-        .when(mockFolioProvider).retrieveResource(
-            argThat((IRequestData data) -> data.getPath().equals(expectedUsersQueryPath)));
+      .when(mockFolioProvider).retrieveResource(usersQueryPath(barcode));
 
     final SessionData sessionData = SessionData.createSession("diku", '|', false, "IBM850");
 
@@ -100,11 +95,6 @@ public class UsersRepositoryTests {
 
     final String username = "leslie";
     final String userId = "4f0e711c-d583-41e0-9555-b62f1725023f";
-    final String expectedUsersQueryPath = "/users?limit=1&query="
-        + Utils.encode("(barcode==" + username
-        + " or externalSystemId==" + username
-        + " or username==" + username + ')');
-
     final String expectedUsersBlQueryPath = "/bl-users/by-id/" + userId;
 
     doReturn(Future.succeededFuture(new FolioResource(userBlResponse,
@@ -114,8 +104,7 @@ public class UsersRepositoryTests {
 
     doReturn(Future.succeededFuture(new FolioResource(userResponse,
         MultiMap.caseInsensitiveMultiMap().add("x-okapi-token", "1234"))))
-        .when(mockFolioProvider).retrieveResource(
-            argThat((IRequestData data) -> data.getPath().equals(expectedUsersQueryPath)));
+      .when(mockFolioProvider).retrieveResource(usersQueryPath(username));
 
     final SessionData sessionData = SessionData.createSession("diku", '|', false, "IBM850");
 
@@ -141,10 +130,6 @@ public class UsersRepositoryTests {
 
     final String username = "leslie";
     final String userId = "4f0e711c-d583-41e0-9555-b62f1725023f";
-    final String expectedUsersQueryPath = "/users?limit=1&query="
-        + Utils.encode("(barcode==" + username
-        + " or externalSystemId==" + username
-        + " or username==" + username + ')');
 
     final String expectedUsersBlQueryPath = "/bl-users/by-id/" + userId;
 
@@ -155,8 +140,7 @@ public class UsersRepositoryTests {
 
     doReturn(Future.succeededFuture(new FolioResource(userResponse,
         MultiMap.caseInsensitiveMultiMap().add("x-okapi-token", "1234"))))
-        .when(mockFolioProvider).retrieveResource(
-        argThat((IRequestData data) -> data.getPath().equals(expectedUsersQueryPath)));
+      .when(mockFolioProvider).retrieveResource(usersQueryPath(username));
 
     final SessionData sessionData = SessionData.createSession("diku", '|', false, "IBM850");
 
@@ -190,10 +174,6 @@ public class UsersRepositoryTests {
 
     final String extSystemId = "4f0e711c-d583-41e0-9555-b62f1725023f";
     final String userId = "4f0e711c-d583-41e0-9555-b62f1725023f";
-    final String expectedUsersQueryPath = "/users?limit=1&query="
-        + Utils.encode("(barcode==" + extSystemId
-        + " or externalSystemId==" + extSystemId
-        + " or username==" + extSystemId + ')');
 
     final String expectedUsersBlQueryPath = "/bl-users/by-id/" + userId;
     IResourceProvider<IRequestData> mockFolioProvider
@@ -206,8 +186,7 @@ public class UsersRepositoryTests {
 
     doReturn(Future.succeededFuture(new FolioResource(userResponse,
         MultiMap.caseInsensitiveMultiMap().add("x-okapi-token", "1234"))))
-        .when(mockFolioProvider).retrieveResource(
-            argThat((IRequestData data) -> data.getPath().equals(expectedUsersQueryPath)));
+      .when(mockFolioProvider).retrieveResource(usersQueryPath(extSystemId));
 
     final SessionData sessionData = SessionData.createSession("diku", '|', false, "IBM850");
 
@@ -227,15 +206,9 @@ public class UsersRepositoryTests {
       @Mock IResourceProvider<IRequestData> mockFolioProvider) {
 
     final String barcode = "1234667";
-    final String userId = "4f0e711c-d583-41e0-9555-b62f1725023f";
-    final String expectedUsersQueryPath = "/users?limit=1&query="
-        + Utils.encode("(barcode==" + barcode
-        + " or externalSystemId==" + barcode
-        + " or username==" + barcode + ')');
 
     doReturn(Future.failedFuture(new NoStackTraceThrowable("Test failure")))
-        .when(mockFolioProvider).retrieveResource(
-            argThat((IRequestData data) -> data.getPath().equals(expectedUsersQueryPath)));
+      .when(mockFolioProvider).retrieveResource(usersQueryPath(barcode));
 
     final SessionData sessionData = SessionData.createSession("diku", '|', false, "IBM850");
 
@@ -260,10 +233,6 @@ public class UsersRepositoryTests {
 
     final String barcode = "997383903573496";
     final String userId = "4f0e711c-d583-41e0-9555-b62f1725023f";
-    final String expectedUsersQueryPath = "/users?limit=1&query="
-        + Utils.encode("(barcode==" + barcode
-        + " or externalSystemId==" + barcode
-        + " or username==" + barcode + ')');
 
     final String expectedUsersBlQueryPath = "/bl-users/by-id/" + userId;
 
@@ -273,8 +242,7 @@ public class UsersRepositoryTests {
 
     doReturn(Future.succeededFuture(new FolioResource(userResponse,
         MultiMap.caseInsensitiveMultiMap().add("x-okapi-token", "1234"))))
-        .when(mockFolioProvider).retrieveResource(
-        argThat((IRequestData data) -> data.getPath().equals(expectedUsersQueryPath)));
+      .when(mockFolioProvider).retrieveResource(usersQueryPath(barcode));
 
     final SessionData sessionData = SessionData.createSession("diku", '|', false, "IBM850");
 
@@ -285,5 +253,13 @@ public class UsersRepositoryTests {
 
           testContext.completeNow();
         })));
+  }
+
+  private IRequestData usersQueryPath(final String searchString) {
+    final String expectedUsersQueryPath = "/users?limit=1&query="
+        + encode("barcode==\"" + searchString + "\""
+        + " or externalSystemId==\"" + searchString + "\""
+        + " or username==\"" + searchString + "\"");
+    return argThat((IRequestData data) -> data.getPath().equals(expectedUsersQueryPath));
   }
 }